行動選択時の指定箇所の範囲表示を実装。（但し移動はまだ未実装）

+++ BattleShip/BattleShip/scr/SetShip.cpp
@@ -15,8 +15,8 @@ bool SetShip::Init()
 int SetShip::Control()
 {
 	m_pStage->ResetSelect();	///<	ステージの選択状態をリセット
-	m_tempX = m_pMouse->GetCursorPosX();	///<	マウス座標の更新
-	m_tempY = m_pMouse->GetCursorPosY();	///<	マウス座標の更新
+	m_tempX = (float)m_pMouse->GetCursorPosX();	///<	マウス座標の更新
+	m_tempY = (float)m_pMouse->GetCursorPosY();	///<	マウス座標の更新
 	m_tempShip = m_pPlayer[m_playerID-1]->GetShip( (ShipObject::_SHIP_TYPE_NUM_)m_ShipCount );
 			
 	if( !m_StateCompFlag )
