+++ BattleShip/Library/scr/FuncRender.cpp
@@ -404,7 +404,7 @@ void CRenderManager::MousePosTransform( long _cursorMovePosX, long _cursorMovePo
 	D3DXMatrixLookAtLH(&matView, &vecEyePt, &vecLookatPt, &vecUpVec);
 	D3DXMatrixPerspectiveFovLH(&matProj, D3DX_PI/4, 1.6f, 0.9f, 100.0f);
 
-	CalcScreenToWorld ( &vecCursorPos, _cursorMovePosX, _cursorMovePosY, 
+	CalcScreenToWorld ( &vecCursorPos, (float)_cursorMovePosX, (float)_cursorMovePosY, 
 						1.f, WIDTH, HEIGHT, &matView, &matProj );
 
 	char msg[256];
